+++ packages/amplify-ui-components/src/common/helpers.ts
@@ -85,7 +85,7 @@ export const isHintValid = field => {
 };
 
 // Required attributes come from https://openid.net/specs/openid-connect-core-1_0.html#StandardClaims
-export const requiredAttributesMap = {
+export const getRequiredAttributesMap = () => ({
   address: {
     label: I18n.get(Translations.ADDRESS_LABEL),
     placeholder: I18n.get(Translations.ADDRESS_PLACEHOLDER),
@@ -128,7 +128,7 @@ export const requiredAttributesMap = {
   },
   given_name: {
     label: I18n.get(Translations.GIVEN_NAME_LABEL),
-    placeholder: I18n.get(Translations.GIVEN_NAME_LABEL),
+    placeholder: I18n.get(Translations.GIVEN_NAME_PLACEHOLDER),
   },
   zoneinfo: {
     label: I18n.get(Translations.ZONEINFO_LABEL),
@@ -154,7 +154,7 @@ export const requiredAttributesMap = {
     label: I18n.get(Translations.NAME_LABEL),
     placeholder: I18n.get(Translations.NAME_PLACEHOLDER),
   },
-};
+});
 
 export function handlePhoneNumberChange(event, phoneNumber: PhoneNumberInterface) {
   const name = event.target.name;