+++ packages/api/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.29](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.28...@aws-amplify/api@3.2.29) (2021-03-25)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [3.2.28](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.27...@aws-amplify/api@3.2.28) (2021-03-18)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [3.2.27](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.26...@aws-amplify/api@3.2.27) (2021-03-12)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [3.2.26](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.25...@aws-amplify/api@3.2.26) (2021-03-08)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [3.2.25](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.24...@aws-amplify/api@3.2.25) (2021-03-03)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
 ## [3.2.24](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@3.2.23...@aws-amplify/api@3.2.24) (2021-02-25)
 
 **Note:** Version bump only for package @aws-amplify/api